refactor: avoid Action Cable "hard" monkey-patching
This PR changes the (historical) way of making Action Cable compatible with AnyCable:

- Extend (via #prepend) instead of reopening classes
- Preserve the original functionality (by using #anycabled? checks where necessary)

Closes #165
