+++ lib/anycable/rails/action_cable_ext/remote_connections.rb
@@ -2,7 +2,7 @@
 
 require "action_cable/remote_connections"
 
-ActionCable::RemoteConnections::RemoteConnection.include(ActionCable::Connection::SerializableIdentification)
+ActionCable::RemoteConnections::RemoteConnection.include(AnyCable::Rails::Connections::SerializableIdentification)
 
 ActionCable::RemoteConnections::RemoteConnection.prepend(Module.new do
   def disconnect(reconnect: true)
